pyproject.toml

[project]
name = "a2a_agents"
version = "0.1.0"
description = "Bootstrapping some useful agents for a2a registry"
authors = [{ name = "Prass, The Nomadic Coder", email = "contact@a2aregistry.org" }]
readme = "README.md"
license = {file = "LICENSE"}
keywords = ["python"]
requires-python = ">=3.11,<4"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Typing :: Typed",
]
dependencies = [
    "pydantic-ai>=0.7.2",
    "fastapi>=0.115.0",
    "modal>=0.65.0",
    "uvicorn>=0.32.0",
    "duckduckgo-search>=6.3.0",
    "GitPython>=3.1.43",
    "httpx>=0.28.0",
    "python-multipart>=0.0.17",
    "python-dotenv>=1.0.1",
    "fasta2a>=0.5.0",
    "a2a>=0.10.0",
]

[dependency-groups]
dev = [
    "ty>=0.0.1a11",
    
    "pytest>=8.4.1",
    "coverage>=7.8.2",
    "ruff>=0.12.0",
    "taskipy>=1.14.1",
    "pre-commit>=4.2.0",
]

[project.urls]
homepage = "https://github.com/prassanna-ravishankar/a2a-agent-bootstrapping"
repository = "https://github.com/prassanna-ravishankar/a2a-agent-bootstrapping.git"
documentation = "https://github.com/prassanna-ravishankar/a2a-agent-bootstrapping"
issues = "https://github.com/prassanna-ravishankar/a2a-agent-bootstrapping/issues"

[build-system]
requires = ["uv_build>=0.7.19,<0.8.0"]
build-backend = "uv_build"

[tool.taskipy.tasks]
format = { cmd = "ruff format ./src/", help = "Format code using ruff" }
lint = { cmd = "ruff check --fix ./src/", help = "Lint code using ruff" }
type = { cmd = "ty check", help = "Typecheck code using ty" }
tests = { cmd = "pytest --verbose --color=yes tests", help = "Run all tests using pytest" }
coverage = { cmd = "coverage run -m pytest tests && coverage report", help = "Generate coverage report" }
run = { cmd = "uv run ./src/a2a_agents/a2a_agents.py", help = "Run main application" }

[tool.ruff]
line-length = 100
indent-width = 4
format.indent-style = "space"
target-version = "py313"
src = ["src", "tests"]

[tool.ruff.lint]
pydocstyle.convention = "google"
fixable = ["ALL"]
select = ["ALL"]
ignore = [
    "FA102", # Flake8-future-annotations
    "F401", # Disable fix for unused imports
    "B904", # Allow raising exceptions without from e
    "PLR0913", # too many arguments
    "CPY", # Flake8-copyright
    "T201", # Debugging print statements
    "ERA", # Eradicate – detects commented-out code
    "BLE001", # Catch-all exceptions
    "D100", # missing docstring in public module
    "D105", # undocumented magic method
    "D107", # undocumented public init
    "D203", # incorrect-blank-line-before-class
    "D205", # line-between-summary-and-description
    "D212", # multi-line-summary-first-line
    "D400", # missing-trailing-period
    "D407", # missing-dashed-underline-after-section
    "S311", # rand-usage
    "EM102", # f-string-exception
    "TRY003", # long-exception-messages
]


[tool.ty]
src.root = "./src"
environment.python = "./.venv"
terminal.output-format = "concise"

[tool.coverage.report]
skip_empty = true
show_missing = true
fail_under = 100

[tool.coverage.run]
branch = true
source = ["src"]
omit = ["_*.py", "__*.py"]

[tool.portray.mkdocs]
site_name = "a2a-agent-bootstrapping"

[tool.portray.mkdocs.theme]
name = "material"
custom_dir = "docs"
palette = {scheme= "slate", primary="deep purple", accent="purple"}

